% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hangman-package.R
\docType{package}
\name{hangman-package}
\alias{hangman-package}
\alias{_PACKAGE}
\title{hangman: Play a game of hangman in the console}
\source{
\strong{hangman} uses a default word list derived from the
\href{http://corpus.leeds.ac.uk/list.html}{Internet corpus} created by the
\href{https://ahc.leeds.ac.uk/centre-for-translation-studies}{Centre for Translation Studies, University of Leeds},
distributed under the
\href{https://creativecommons.org/licenses/by/2.5/}{Creative Commons Attribution 2.5 Generic}
license.

This word list is passed through the
\href{https://www.cs.cmu.edu/~biglou/resources/}{Offensive/Profane Word List}
created by \href{https://www.cs.cmu.edu/~biglou/}{Luis Von Ahn} to filter out
potentially offensive words.
}
\description{
Creates an interactive game in the console. Try to reveal a
  randomly selected word by guessing letters. If you make six incorrect
  guesses, you lose. Includes a list of 10,000 common English words with more
  than three letters.
}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Centre for Translation Studies, University of Leeds [copyright holder]
  \item Luis Von Ahn [copyright holder]
}

}
\keyword{internal}
