% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_hangman.R
\name{play_hangman}
\alias{play_hangman}
\alias{hangman}
\title{Play a game of hangman in the console}
\source{
The default word list is derived from the
\href{http://corpus.leeds.ac.uk/list.html}{Internet corpus} created by the
\href{https://ahc.leeds.ac.uk/centre-for-translation-studies}{Centre for Translation Studies, University of Leeds},
distributed under the
\href{https://creativecommons.org/licenses/by/2.5/}{Creative Commons Attribution 2.5 Generic}
license.

This word list is passed through the
\href{https://www.cs.cmu.edu/~biglou/resources/}{Offensive/Profane Word List}
created by \href{https://www.cs.cmu.edu/~biglou/}{Luis Von Ahn} to filter out
potentially unwanted words.
}
\usage{
play_hangman(word_list = hangman_word_list)
}
\arguments{
\item{word_list}{A character vector of words used to play the hangman game.
Defaults to a list of 10,000 common English words with four or
more letters.}
}
\value{

}
\description{
Play a game of hangman in the console
}
\examples{

play_hangman()

play_hangman("test")

cars <- unique(gsub(" .*", "", rownames(mtcars)))
play_hangman(cars)
}
