\name{hangman-package}
\alias{hangman-package}
\alias{hangman}
\docType{package}
\title{
hangman game
}
\description{
hangman game
}
\details{
\tabular{ll}{
Package: \tab hangman\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-01\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tyler Rinker

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
